feat: add snapshot status indicator
