--- src/components/Icon/SnapshotFailing.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-function SnapshotFailing() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 14 14" fill="none">
-      <circle cx="7" cy="7" r="7" fill="#D80027" />
-      <rect x="3" y="6" width="8" height="2" fill="white" />
-    </svg>
-  )
-}
-
-export default SnapshotFailing
