--- src/components/Icon/SnapshotOk.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-function SnapshotOk() {
-  return (
-    <svg xmlns="http://www.w3.org/2000/svg" width="14" height="14" viewBox="0 0 14 14" fill="none">
-      <circle cx="7" cy="7" r="7" fill="#328500" />
-      <path d="M3 7L6 10L11 5" stroke="white" strokeWidth="2" strokeLinecap="round" />
-    </svg>
-  )
-}
-
-export default SnapshotOk
