+++ src/components/Layout/Layout.tsx
@@ -9,6 +9,7 @@ import { Navbar, NavbarProps } from 'decentraland-ui/dist/components/Navbar/Navb
 import type { PageProps } from 'gatsby'
 
 import { isProjectPath } from '../../utils/locations'
+import SnapshotStatus from '../Debug/SnapshotStatus'
 import WalletSelectorModal from '../Modal/WalletSelectorModal'
 import WrongNetworkModal from '../Modal/WrongNetworkModal'
 
@@ -48,6 +49,7 @@ export default function Layout({ children, ...props }: LayoutProps) {
 
   return (
     <>
+      <SnapshotStatus />
       <Navbar
         activePage="dao"
         onClickMenuOption={handleClickMenuOption}
