+++ src/pages/debug.tsx
@@ -10,7 +10,7 @@ import BudgetsUpdate from '../components/Debug/BudgetsUpdate'
 import EnvStatus from '../components/Debug/EnvStatus'
 import ErrorReporting from '../components/Debug/ErrorReporting'
 import HttpStatus from '../components/Debug/HttpStatus'
-import SnapshotStatus from '../components/Debug/SnapshotStatus'
+import Snapshot from '../components/Debug/Snapshot'
 import TriggerFunction from '../components/Debug/TriggerFunction'
 import LogIn from '../components/Layout/LogIn'
 import Navigation, { NavigationTab } from '../components/Layout/Navigation'
@@ -34,7 +34,7 @@ export default function DebugPage() {
       <BudgetsUpdate className="DebugPage__Section" />
       <HttpStatus className="DebugPage__Section" />
       <BadgesAdmin className="DebugPage__Section" />
-      <SnapshotStatus className="DebugPage__Section" />
+      <Snapshot className="DebugPage__Section" />
       <EnvStatus className="DebugPage__Section" />
       <TriggerFunction className="DebugPage__Section" />
       <ErrorReporting className="DebugPage__Section" />
